FDG-10500 add helper functions
