+++ src/components/data-preview/data-preview-filter-section/data-preview-table-filters/data-preview-filter-helper.js
@@ -8,3 +8,37 @@ export const getDaysArray = (start, end) => {
   }
   return arr;
 };
+
+export const basePreset = [{ label: 'All', key: 'all', years: null }];
+export const possiblePresets = [
+  { label: '1 Year', key: '1yr', years: 1 },
+  { label: '5 Years', key: '5yr', years: 5 },
+  { label: '10 Years', key: '10yr', years: 10 },
+];
+export const customPreset = { label: 'Custom', key: 'custom', years: null };
+// If a data table has less than 5 years of data, we need to find the next best option to select
+// by default.
+export const fallbackPresets = ['1yr', 'current', 'all'];
+
+export const createFilterConfigs = (fields, datePreset, selectedTable) => {
+  const fieldsConfig = [...fields];
+  fieldsConfig.forEach(field => {
+    if (field.dataType === 'DATE') {
+      if (!field?.pendingStartDate) {
+        field.pendingStartDate = null;
+      }
+      if (!field?.pendingEndDate) {
+        field.pendingEndDate = null;
+      }
+      if (field.columnName === selectedTable?.dateField && datePreset) {
+        field.defaultStartDate = datePreset?.from;
+        field.defaultEndDate = datePreset?.to;
+      }
+    } else {
+      if (!field?.pendingValue) {
+        field.pendingValue = '';
+      }
+    }
+  });
+  return fieldsConfig;
+};
